% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_parametric_model.R
\name{all_parametric_model}
\alias{all_parametric_model}
\title{Multiple parametric models for estimating species richness with various cutoff values}
\usage{
all_parametric_model(
  input_data,
  parallel = F,
  tau_range = NULL,
  control = list(ncores = ceiling(detectCores()/2))
)
}
\arguments{
\item{input_data}{An input type that can be processed by convert()}

\item{parallel}{A logic scalar that decides whether the richness estimates are computed with parallelization.
Currently an error is returned for Windows.}

\item{tau_range}{A two-dimensional vector of positive intergers specifing the range of the cutoffs you'd like to vary within}

\item{control}{A list containing an integer \code{ncores} that specifies the number of cores to use
in parallelization when \code{parallel == T}}
}
\value{
A data frame displaying the point estimates, standard errors, AICc, GOF0 and GOF5 for different
parametric models and cutoffs.
}
\description{
This function generates the species richness estimates with Poisson model, geometric model,
two-component geometric mixture model and three-component geometric model.
}
\examples{
library(breakaway)
data(apples)
all_parametric_model(apples)
}
