% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_models.R
\name{geometric_model}
\alias{geometric_model}
\title{Geometric model for estimating species richness}
\usage{
geometric_model(input_data, cutoff = 10)
}
\arguments{
\item{input_data}{An input type that can be processed by convert()}

\item{cutoff}{Maximal frequency count of the data used to estimating the species richness. Default 10.}
}
\value{
An object of class \code{alpha_estimate}
}
\description{
This function implements the species richness estimation with single-exponential mixed Poisson distribution (or geometric distribution)
}
\examples{
library(breakaway)
data(apples)
geometric_model(apples, cutoff = 20)
}
