% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_parametric_model.R
\name{select_best_models}
\alias{select_best_models}
\title{"Best" parametric models for estimating species richness with various cutoff values}
\usage{
select_best_models(
  input_data,
  parallel = F,
  tau_range = NULL,
  control = list(ncores = ceiling(detectCores()/2))
)
}
\arguments{
\item{input_data}{An input type that can be processed by convert()}

\item{parallel}{A logic scalar that decides whether the richness estimates are computed with parallelization.
Currently an error is returned for Windows.}

\item{control}{A list containing an integer \code{ncores} that specifies the number of cores to use
in parallelization when \code{parallel == T}}
}
\value{
A data frame displaying the point estimates, standard errors, AICc, GOF0 and GOF5 for different
parametric models and cutoffs.
}
\description{
This function gives the "best" models among Poisson model, geometric model,
two-component geometric mixture model and three-component geometric model.
}
\examples{
library(breakaway)
data(apples)
select_best_models(apples)

data(hawaii)
select_best_models(hawaii)
}
