% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three_mixed_exp.R
\name{three_geometric_model}
\alias{three_geometric_model}
\title{Three-component geometric mixture model for estimating species richness}
\usage{
three_geometric_model(input_data, cutoff = 10, ...)
}
\arguments{
\item{input_data}{An input type that can be processed by convert()}

\item{cutoff}{Maximal frequency count of the data used to estimating the species richness. Default 10.}
}
\value{
An object of class \code{alpha_estimate}
}
\description{
This function implements the species richness estimation with a three-component mixture
exponential-mixed Poisson model (or three-component mixture exponential-mixed geometric model)
for estimating species richness
}
\examples{
library(breakaway)
data(apples)
three_geometric_model(apples, cutoff = 50)
}
