% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_mixed_exp.R
\name{two_mixed_exp_lld}
\alias{two_mixed_exp_lld}
\title{Internal functions for estimating species richness with a two-component geomic mixture model}
\usage{
two_mixed_exp_lld(input_data, t1, t2, u)
}
\description{
Internal functions for estimating species richness with a two-component geomic mixture model
}
